requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==24.2.0
base58==2.1.1
bech32==1.2.0
bitarray==3.0.0
bitcoinlib==0.7.0
blessed==1.20.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ckzg==2.0.1
coincurve==20.0.0
cytoolz==1.0.0
ecdsa==0.19.0
editor==1.6.6
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
exceptiongroup==1.2.2
fastecdsa==2.3.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inquirer==3.4.0
multidict==6.1.0
numpy==2.1.3
parsimonious==0.10.0
prompt-toolkit==1.0.14
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
pyunormalize==16.0.0
readchar==4.2.1
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
runs==1.2.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
toolz==1.0.0
tronpy==0.5.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
web3==7.6.0
websockets==13.1
xmod==1.8.1
yarl==1.18.3
